[#91] refactor: split v1_iks tests into per-concern package
Replace the 700-line tests.py with a tests/ package: a shared
BaseIKSTestCase and one file per concern (seeder, download command,
stats, series, indicators, download trigger, aggregations).
