+++ backend/api/v1/v1_iks/tests/tests_iks_stats_endpoint.py
@@ -1,0 +1,106 @@
+from datetime import datetime
+from django.utils import timezone
+from rest_framework import status
+from api.v1.v1_iks.models import KoboData, IKSIndicator, IKSValue
+from .base import BaseIKSTestCase
+
+
+class IKSStatsEndpointTests(BaseIKSTestCase):
+
+    def test_iks_stats_endpoint(self):
+        """Test GET /api/v1/iks/{administration_id}/stats API."""
+        indicator = IKSIndicator.objects.create(
+            kobo_form=self.form, name="test_indicator_drought"
+        )
+        KoboData.objects.create(
+            form=self.form,
+            kobo_id=123,
+            submission_time=timezone.now(),
+            raw_data={"_validation_status": {"uid": "val-1"}},
+        )
+        IKSValue.objects.create(
+            kobo_id=123,
+            administration=self.admin_area,
+            iks_indicator=indicator,
+            value="observed",
+        )
+
+        response = self.client.get(f"/api/v1/iks/{self.admin_area.id}/stats")
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        self.assertEqual(response.json()["total_reports_received"], 1)
+        self.assertEqual(response.json()["total_months_drought"], 1)
+
+    def test_iks_stats_endpoint_empty(self):
+        """
+        Test GET /api/v1/iks/{administration_id}/stats API with empty database.
+        """
+        response = self.client.get(f"/api/v1/iks/{self.admin_area.id}/stats")
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        self.assertEqual(response.json()["total_reports_received"], 0)
+        self.assertEqual(response.json()["total_months_drought"], 0)
+        self.assertEqual(response.json()["validation_rate_percentage"], 0.0)
+
+    def test_iks_stats_endpoint_anonymous(self):
+        """
+        Test GET /api/v1/iks/{administration_id}/stats
+        API guards authentication.
+        """
+        self.client.force_authenticate(user=None)
+        response = self.client.get(f"/api/v1/iks/{self.admin_area.id}/stats")
+        self.assertEqual(response.status_code, status.HTTP_401_UNAUTHORIZED)
+
+    def test_iks_stats_date_filters(self):
+        """Test GET /api/v1/iks/{administration_id}/stats date filtering."""
+        indicator = IKSIndicator.objects.create(
+            kobo_form=self.form, name="test_indicator_drought"
+        )
+
+        # 1. Submission inside range (May 2026)
+        KoboData.objects.create(
+            form=self.form,
+            kobo_id=100,
+            submission_time=timezone.make_aware(
+                datetime(2026, 5, 10, 12, 0, 0)
+            ),
+            raw_data={"_validation_status": {"uid": "val-100"}},
+        )
+        val_inside = IKSValue.objects.create(
+            kobo_id=100,
+            administration=self.admin_area,
+            iks_indicator=indicator,
+            value="observed",
+        )
+        # Manually backdate created timestamp to test drought month calculation
+        val_inside.created = timezone.make_aware(
+            datetime(2026, 5, 10, 12, 0, 0)
+        )
+        val_inside.save()
+
+        # 2. Submission outside range (Jan 2026)
+        KoboData.objects.create(
+            form=self.form,
+            kobo_id=200,
+            submission_time=timezone.make_aware(
+                datetime(2026, 1, 15, 12, 0, 0)
+            ),
+            raw_data={"_validation_status": {"uid": "val-200"}},
+        )
+        val_outside = IKSValue.objects.create(
+            kobo_id=200,
+            administration=self.admin_area,
+            iks_indicator=indicator,
+            value="observed",
+        )
+        val_outside.created = timezone.make_aware(
+            datetime(2026, 1, 15, 12, 0, 0)
+        )
+        val_outside.save()
+
+        # Request with date filters matching May 2026
+        response = self.client.get(
+            f"/api/v1/iks/{self.admin_area.id}/stats",
+            {"start_date": "2026-05-01", "end_date": "2026-05-31"},
+        )
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        self.assertEqual(response.json()["total_reports_received"], 1)
+        self.assertEqual(response.json()["total_months_drought"], 1)
